+++ tests/ActiveDataProviderTest.php
@@ -4,6 +4,8 @@ namespace yiiunit\extensions\sphinx;
 
 use yii\sphinx\ActiveDataProvider;
 use yii\sphinx\Query;
+use Yii;
+use yii\web\Request;
 use yiiunit\extensions\sphinx\data\ar\ActiveRecord;
 use yiiunit\extensions\sphinx\data\ar\ArticleIndex;
 
@@ -30,7 +32,7 @@ class ActiveDataProviderTest extends TestCase
             'db' => $this->getConnection(),
         ]);
         $models = $provider->getModels();
-        $this->assertEquals(2, count($models));
+        $this->assertEquals(20, count($models));
 
         $provider = new ActiveDataProvider([
             'query' => $query,
@@ -49,10 +51,10 @@ class ActiveDataProviderTest extends TestCase
             'query' => ArticleIndex::find()->orderBy('id ASC'),
         ]);
         $models = $provider->getModels();
-        $this->assertEquals(2, count($models));
+        $this->assertEquals(20, count($models));
         $this->assertTrue($models[0] instanceof ArticleIndex);
         $this->assertTrue($models[1] instanceof ArticleIndex);
-        $this->assertEquals([1, 2], $provider->getKeys());
+        $this->assertEquals([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20], $provider->getKeys());
 
         $provider = new ActiveDataProvider([
             'query' => ArticleIndex::find(),
@@ -80,8 +82,8 @@ class ActiveDataProviderTest extends TestCase
             'db' => $this->getConnection(),
         ]);
         $models = $provider->getModels();
-        $this->assertEquals(2, count($models));
-        $this->assertEquals(2, count($provider->getFacet('author_id')));
+        $this->assertEquals(20, count($models));
+        $this->assertEquals(20, count($provider->getFacet('author_id')));
     }
 
     /**
@@ -102,6 +104,82 @@ class ActiveDataProviderTest extends TestCase
         ]);
         $models = $provider->getModels();
         $this->assertEquals(1, count($models));
-        $this->assertEquals(2, $provider->getTotalCount());
+        $this->assertEquals(1002, $provider->getTotalCount());
+    }
+
+    /**
+     * @depends testTotalCountFromMeta
+     *
+     * @see https://github.com/yiisoft/yii2-sphinx/issues/11
+     */
+    public function testAutoAdjustPagination()
+    {
+        $request = new Request();
+        $request->setQueryParams(['page' => 2]);
+        Yii::$app->set('request', $request);
+
+        $query = new Query();
+        $query->from('yii2_test_article_index');
+        $query->orderBy(['id' => SORT_ASC]);
+        $query->showMeta(true);
+
+        $provider = new ActiveDataProvider([
+            'query' => $query,
+            'db' => $this->getConnection(),
+            'pagination' => [
+                'pageSize' => 1,
+            ]
+        ]);
+        $models = $provider->getModels();
+        $this->assertEquals(2, $models[0]['id']);
+    }
+
+    /**
+     * @depends testAutoAdjustPagination
+     *
+     * @see https://github.com/yiisoft/yii2-sphinx/issues/12
+     */
+    public function testAutoAdjustMaxMatches()
+    {
+        $request = new Request();
+        $request->setQueryParams(['page' => 99999]);
+        Yii::$app->set('request', $request);
+
+        $query = new Query();
+        $query->from('yii2_test_article_index');
+        $query->orderBy(['id' => SORT_ASC]);
+
+        $provider = new ActiveDataProvider([
+            'query' => $query,
+            'db' => $this->getConnection(),
+            'pagination' => [
+                'pageSize' => 100,
+                'validatePage' => false,
+            ]
+        ]);
+        $models = $provider->getModels();
+        $this->assertEmpty($models); // no exception
+    }
+
+    public function testMatch()
+    {
+        $query = new Query();
+        $query->from('yii2_test_article_index');
+        $query->match('Repeated');
+
+        $provider = new ActiveDataProvider([
+            'query' => $query,
+            'db' => $this->getConnection(),
+        ]);
+
+        $this->assertEquals(1002, $provider->getTotalCount());
+
+        $query->match('Excepturi');
+        $provider = new ActiveDataProvider([
+            'query' => $query,
+            'db' => $this->getConnection(),
+        ]);
+
+        $this->assertEquals(29, $provider->getTotalCount());
     }
 }